Refactored card tests to use shared list of cards that's populated once
Drastically speeds up these tests, and allowed me to break them up into discrete chunks the way they should be.

Closes #48.
